Add code to all AMA flows
