+++ client/app/intake/actions/ama.js
@@ -99,3 +99,65 @@ export const completeIntake = (intakeId, intakeData) => (dispatch) => {
       }
     );
 };
+
+export const setClaimantNotVeteran = (claimantNotVeteran) => ({
+  type: ACTIONS.SET_CLAIMANT_NOT_VETERAN,
+  payload: {
+    claimantNotVeteran
+  }
+});
+
+export const setClaimant = (claimant) => ({
+  type: ACTIONS.SET_CLAIMANT,
+  payload: {
+    claimant
+  }
+});
+
+export const setIssueSelected = (profileDate, issueId, isSelected) => ({
+  type: ACTIONS.SET_ISSUE_SELECTED,
+  payload: {
+    profileDate,
+    issueId,
+    isSelected
+  },
+  meta: {
+    analytics: {
+      label: isSelected ? 'selected' : 'de-selected'
+    }
+  }
+});
+
+export const addNonRatedIssue = (nonRatedIssues) => ({
+  type: ACTIONS.ADD_NON_RATED_ISSUE,
+  payload: {
+    nonRatedIssues
+  },
+  meta: { analytics }
+});
+
+export const setIssueCategory = (issueId, category) => ({
+  type: ACTIONS.SET_ISSUE_CATEGORY,
+  payload: {
+    issueId,
+    category
+  },
+  meta: {
+    analytics: {
+      label: category
+    }
+  }
+});
+
+export const setIssueDescription = (issueId, description) => ({
+  type: ACTIONS.SET_ISSUE_DESCRIPTION,
+  payload: {
+    issueId,
+    description
+  },
+  meta: {
+    analytics: {
+      label: description
+    }
+  }
+});
