+++ client/app/intake/actions/common.js
@@ -91,20 +91,6 @@ export const setReceiptDate = (receiptDate) => ({
   }
 });
 
-export const setClaimantNotVeteran = (claimantNotVeteran) => ({
-  type: ACTIONS.SET_CLAIMANT_NOT_VETERAN,
-  payload: {
-    claimantNotVeteran
-  }
-});
-
-export const setClaimant = (claimant) => ({
-  type: ACTIONS.SET_CLAIMANT,
-  payload: {
-    claimant
-  }
-});
-
 export const setOptionSelected = (optionSelected) => ({
   type: ACTIONS.SET_OPTION_SELECTED,
   payload: {
@@ -113,54 +99,6 @@ export const setOptionSelected = (optionSelected) => ({
   meta: {
     analytics: {
       label: optionSelected
-    }
-  }
-});
-
-export const setIssueSelected = (profileDate, issueId, isSelected) => ({
-  type: ACTIONS.SET_ISSUE_SELECTED,
-  payload: {
-    profileDate,
-    issueId,
-    isSelected
-  },
-  meta: {
-    analytics: {
-      label: isSelected ? 'selected' : 'de-selected'
-    }
-  }
-});
-
-export const addNonRatedIssue = (nonRatedIssues) => ({
-  type: ACTIONS.ADD_NON_RATED_ISSUE,
-  payload: {
-    nonRatedIssues
-  },
-  meta: { analytics }
-});
-
-export const setIssueCategory = (issueId, category) => ({
-  type: ACTIONS.SET_ISSUE_CATEGORY,
-  payload: {
-    issueId,
-    category
-  },
-  meta: {
-    analytics: {
-      label: category
-    }
-  }
-});
-
-export const setIssueDescription = (issueId, description) => ({
-  type: ACTIONS.SET_ISSUE_DESCRIPTION,
-  payload: {
-    issueId,
-    description
-  },
-  meta: {
-    analytics: {
-      label: description
     }
   }
 });
