+++ client/app/intake/pages/appeal/finish.jsx
@@ -7,8 +7,7 @@ import CancelButton from '../../components/CancelButton';
 import NonRatedIssuesUnconnected from '../../components/NonRatedIssues';
 import RatedIssuesUnconnected from '../../components/RatedIssues';
 import IssueCounter from '../../components/IssueCounter';
-import { setIssueSelected, addNonRatedIssue, setIssueCategory, setIssueDescription } from '../../actions/common';
-import { completeIntake } from '../../actions/ama';
+import { completeIntake, setIssueSelected, addNonRatedIssue, setIssueCategory, setIssueDescription } from '../../actions/ama';
 import { REQUEST_STATE, PAGE_PATHS, INTAKE_STATES } from '../../constants';
 import { getIntakeStatus } from '../../selectors';
 import CompleteIntakeErrorAlert from '../../components/CompleteIntakeErrorAlert';
