+++ client/app/intake/pages/appeal/review.jsx
@@ -8,8 +8,8 @@ import { Redirect } from 'react-router-dom';
 import Button from '../../../components/Button';
 import SelectClaimant from '../../components/SelectClaimant';
 import { setDocketType } from '../../actions/appeal';
-import { submitReview } from '../../actions/ama';
-import { setReceiptDate, setClaimantNotVeteran, setClaimant } from '../../actions/common';
+import { submitReview, setClaimantNotVeteran, setClaimant } from '../../actions/ama';
+import { setReceiptDate } from '../../actions/common';
 import { REQUEST_STATE, PAGE_PATHS, INTAKE_STATES } from '../../constants';
 import { getIntakeStatus } from '../../selectors';
