+++ client/app/intake/pages/higherLevelReview/review.jsx
@@ -8,8 +8,8 @@ import CancelButton from '../../components/CancelButton';
 import Button from '../../../components/Button';
 import SelectClaimant from '../../components/SelectClaimant';
 import { setInformalConference, setSameOffice } from '../../actions/higherLevelReview';
-import { submitReview } from '../../actions/ama';
-import { setReceiptDate, setClaimantNotVeteran, setClaimant } from '../../actions/common';
+import { submitReview, setClaimantNotVeteran, setClaimant } from '../../actions/ama';
+import { setReceiptDate } from '../../actions/common';
 import { REQUEST_STATE, PAGE_PATHS, INTAKE_STATES, BOOLEAN_RADIO_OPTIONS } from '../../constants';
 import { getIntakeStatus } from '../../selectors';
