+++ spec/feature/intake/supplemental_claim_spec.rb
@@ -135,7 +135,7 @@ RSpec.feature "Supplemental Claim Intake" do
     expect(page).to have_content("Left knee granted")
     expect(page).to_not have_content("Untimely rating issue 1")
     expect(page).to have_button("Establish EP", disabled: true)
-    expect(page).to have_content("0 rated issues")
+    expect(page).to have_content("0 issues")
 
     supplemental_claim = SupplementalClaim.find_by(veteran_file_number: "12341234")
 
@@ -147,11 +147,11 @@ RSpec.feature "Supplemental Claim Intake" do
     intake = Intake.find_by(veteran_file_number: "12341234")
 
     find("label", text: "PTSD denied").click
-    expect(page).to have_content("1 rated issue")
+    expect(page).to have_content("1 issue")
     find("label", text: "Left knee granted").click
-    expect(page).to have_content("2 rated issues")
+    expect(page).to have_content("2 issues")
     find("label", text: "Left knee granted").click
-    expect(page).to have_content("1 rated issue")
+    expect(page).to have_content("1 issue")
 
     safe_click "#button-add-issue"
 
@@ -160,7 +160,11 @@ RSpec.feature "Supplemental Claim Intake" do
     fill_in "Issue category", with: "Active Duty Adjustments"
     find("#issue-category").send_keys :enter
 
+    expect(page).to have_content("1 issue")
+
     fill_in "Issue description", with: "Description for Active Duty Adjustments"
+
+    expect(page).to have_content("2 issues")
 
     safe_click "#button-finish-intake"
